faet : 투표 작업중

+++ app/src/main/java/com/hgh/na_o_man/presentation/ui/detail/vote/VoteMainScreen.kt
@@ -242,7 +242,7 @@ fun VoteListScreen(
                 getVoteList(
                     title = vote.title,
                     images = vote.agendaPhotoInfoList.map { it.url },
-                    voteId = vote.agendaId.toLong(),
+                    voteId = vote.agendaId,
                     onClick = { agendaId ->
                         viewModel.setEvent(VoteMainContract.VoteMainEvent.OnAgendaItemClicked(agendaId))
                     }
